SpreadsheetStorageXXX was SpreadsheetTerminalStorageXXX
- Only classes renamed, no actual statements or logic changed.
